Split generic bound type arg tests
